backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.13
alembic==1.14.0
pydantic-settings==2.7.0
python-multipart==0.0.19
PyJWT==2.10.1
joblib==1.4.2
python-dateutil==2.9.0.post0
torch==2.7.0; python_version >= "3.10"
scikit-learn==1.5.2; python_version < "3.14"
pandas==2.2.3; python_version < "3.14"
